[cxx-interop] [NFC] Move safety heuristics logic out of ClangImporter.cpp
